Fix 4 compiler bugs: ORD(BYTE), SHORT chain, INCL/EXCL root, ReadInt wire format
- sema: ORD now accepts BYTE (CP §10.2 BlackBox extension), unblocking
  TextRulers.cp and the ~15 tests that import it transitively
- IR lower: SHORT type chain extended with IrType::I16 → U8, completing
  the LongInt→Integer→IntShort→ShortInt→Byte narrowing ladder
- sema: INCL/EXCL builtin resolves root variable name from
  designator.base.module when the parser emits a QualIdent (e.g. a.opts
  parsed as module="a", name="opts") — fixes TextRulers SET operations
- test fixture: TextModelsHello.odc now writes the run-list piece length
  as i16 (2 bytes), matching BlackBox's ReadInt wire format

474/490 newcp-tests pass; remaining 16 are under active investigation
(TextViews pointer-cast IR bug + Tour.odc decode path).

+++ NewCP/tests/newcp-tests/src/lib.rs
@@ -3254,7 +3254,7 @@ mod tests {
         body.extend_from_slice(&[0u8, 1, 2, 3, 4, 5]); // super-versions
         body.extend_from_slice(&6i32.to_le_bytes());   // run-list length
         body.push(1);                                   // ano = 1
-        body.extend_from_slice(&5i32.to_le_bytes());   // len = 5
+        body.extend_from_slice(&5i16.to_le_bytes());   // len = 5 (ReadInt = 2-byte i16 BB wire)
         body.push(0xFF);                                // terminator
         body.extend_from_slice(b"Hello");
         match stage_synthetic_odc(
